Add files via upload
The game. But a little bit broken. I added a thing where if you are found by the Gru, it makes you do its math homework and if you fail to do it correctly it throws you back in the maze and restarts the game. But at the moment, when it throws you into the maze it skips straight to the math homework part.